Fix error with already joined session
This fixes a situation where the login fails with the error "Already
joined to a session" but the user is kept on the login page.

This can happen if one joins a session and navigates back with the
browser afterwards.  This leads the user to the login page ("/login").
The login will always fail with "Already joined to a session".  But the
user stays on the login page ("/login").

This has been fixed by navigating to the main page ("/") in such cases.